Move migrations to source tree
Users can now run `queue-support-migrate` when installing this library both as a Python package and as a Debian package.

When installing as a Debian package, migrations were shipped and the directory was configured in `alembic.debian.ini`. However, when installing outside Debian, migrations were NOT shipped.

To resolve this, the migrations directory is now included in the source tree.

However, since we can't know the absolute path to said migrations directory (after all, the library could be installed system-wide, or in a venv, or in any other situation), we can't hardcode the path to said directory in `alembic.ini`.

Although Alembic does provide the `%(here)s` primitive to resolve that, I opted for dropping the `alembic.ini` file altogether: it contains only default settings, and shipping it requires many -redundant- files and logic to exist.

Note that a single `alembic.ini` still exists in the project root: that one is not meant for migrating, but for creating migrations / local development.