Allow group to read and write to database
This is needed for ClusterConfigurationManager, as it does not run under the `queue-support` user, , but does read from and write to the database.

Note: in theory, 0750 permissions would suffice for the database dir. However, SQLite requires the user to be able to write to the parent directory of a database.