Ignore extraneous config options
This package is used as a library in other projects. In such projects, for local development, an `.env` file may exist with settings for both `QueueSupport` and the project at hand.

This commit fixes Pydantic complaining about unknown settings - corresponding to the project using this library - by ignoring them, as was the default behaviour in Pydantic v1.